Refined the Post page
still not done, requires (a) removal of the * by the required fields on
the "I'll set the details, and let others join me" option. and (b) need
to have a radio box that toggles between I want to do an activity (i)
right away or (ii) at a set time. The boxes for a set time are not
created and the code for toggling the code is not created either.