first wave for skinning
includes backup files. new logo. few things broken: the
bootstrap.min.js tha’s in the responsive template is commented out in
the base.html - it conflicts with the same file in the boiler plate.
Using the newer one kills the menu. This also breaks the links… the
directory name as specified in the routes file (e.g., act) is now
expeced to serve the image and other files. Also, the BaseHandler
cannot be found…